backend/requirements.txt
fastapi==0.104.1
uvicorn==0.34.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
aiosqlite==0.19.0
starlette==0.46.1
mcp==1.4.1
httpx==0.28.1
httpx-sse==0.4.0
pydantic==2.10.6
pydantic-settings==2.8.1
sse-starlette==2.2.1
python-dotenv==1.0.0
